+++ plugin-document/src/main/java/nl/xillio/xill/plugins/mongodb/constructs/RegexConstruct.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright (C) 2014 Xillio (support@xillio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nl.xillio.xill.plugins.mongodb.constructs;
+
+import nl.xillio.xill.api.components.MetaExpression;
+import nl.xillio.xill.api.construct.Argument;
+import nl.xillio.xill.api.construct.Construct;
+import nl.xillio.xill.api.construct.ConstructContext;
+import nl.xillio.xill.api.construct.ConstructProcessor;
+import nl.xillio.xill.plugins.mongodb.data.MongoRegex;
+
+import java.util.regex.Pattern;
+
+/**
+ * Creates a Mongo regex object.
+ *
+ * @author Xavier
+ */
+public class RegexConstruct extends Construct {
+
+    @Override
+    public ConstructProcessor prepareProcess(ConstructContext context) {
+        return new ConstructProcessor(this::process, new Argument("pattern", ATOMIC));
+    }
+
+    private MetaExpression process(MetaExpression pattern) {
+        String regex = pattern.getStringValue();
+        MetaExpression result = fromValue(regex);
+        result.storeMeta(new MongoRegex(Pattern.compile(regex)));
+        return result;
+    }
+}