+++ plugin-document/src/test/java/nl/xillio/xill/plugins/mongodb/services/FindOneAndDeleteOptionsFactoryTest.java
@@ -18,10 +18,7 @@ package nl.xillio.xill.plugins.mongodb.services;
 import com.mongodb.client.model.FindOneAndDeleteOptions;
 import nl.xillio.xill.TestUtils;
 import nl.xillio.xill.api.components.MetaExpression;
-import nl.xillio.xill.plugins.mongodb.services.serializers.BinarySerializer;
-import nl.xillio.xill.plugins.mongodb.services.serializers.MongoSerializer;
-import nl.xillio.xill.plugins.mongodb.services.serializers.ObjectIdSerializer;
-import nl.xillio.xill.plugins.mongodb.services.serializers.UUIDSerializer;
+import nl.xillio.xill.plugins.mongodb.services.serializers.*;
 import org.testng.annotations.Test;
 
 import java.util.LinkedHashMap;
@@ -32,7 +29,11 @@ import static org.testng.Assert.*;
 public class FindOneAndDeleteOptionsFactoryTest extends TestUtils {
     @Test
     public void testBuildOptions() {
-        MongoConverter mongoConverter = new MongoConverter(new MongoSerializer(new ObjectIdSerializer(), new UUIDSerializer(), new BinarySerializer()));
+        MongoConverter mongoConverter = new MongoConverter(new MongoSerializer(
+                new ObjectIdSerializer(),
+                new UUIDSerializer(),
+                new MongoRegexSerializer(),
+                new BinarySerializer()));
         FindOneAndDeleteOptionsFactory findOneAndDeleteOptionsFactory = new FindOneAndDeleteOptionsFactory(mongoConverter);
         LinkedHashMap<String, MetaExpression> object = new LinkedHashMap<>();
 
@@ -58,7 +59,6 @@ public class FindOneAndDeleteOptionsFactoryTest extends TestUtils {
         assertEquals(options.getMaxTime(TimeUnit.MILLISECONDS), 0);
         object.put("maxTime", fromValue(2000));
         options = findOneAndDeleteOptionsFactory.build(fromValue(object));
-        assertNotNull(options.getMaxTime(TimeUnit.MILLISECONDS));
         assertEquals(MetaExpression.parseObject(options.getMaxTime(TimeUnit.MILLISECONDS)), object.get("maxTime"));
     }
 }